requirements.txt
Django==5.1.5
djangorestframework==3.15.2
dj-rest-auth==7.0.1
django-allauth==65.4.1
